Speed up MPFIT fitting with dtype fix and xp_dispatch for halo/baryon models
Two changes to eliminate JAX dispatch overhead on the MPFIT path:

1. Fix float32 dtype mismatch in convolve: Change cube_final in
   simulate_cube from default float32 to explicit float64, eliminating
   expensive float32->float64 upcasting in scipy fftconvolve when the
   beam kernel is float64. This reduces convolve from ~255 ms to ~46 ms
   per iteration (2D).

2. Add xp_dispatch to halo enclosed_mass and evaluate methods (NFW,
   TwoPowerHalo, Burkert, Einasto, DekelZhao, LinearNFW) and baryon
   helper functions (menc_exp_disk, vcirc_exp_disk, v_circular). On the
   MPFIT path these receive numpy arrays and use np.* ops (zero dispatch
   overhead); on the JAX-loss path they receive JAX tracers and use
   jnp.* ops. Also restore _safe_gammaincinv usage in baryons.py and
   kinematic_options.py that was lost during a git checkout.

Results (per-iteration, CPU):
  2D mpfit_chisq: 673 ms -> 145 ms (4.6x, now faster than main's 221 ms)
  1D mpfit_chisq: 584 ms -> 230 ms (2.5x, faster than main's 274 ms)
  3D mpfit_chisq: 578 ms -> 362 ms (1.6x, matches main's 364 ms)

+++ dysmalpy/models/model_set.py
@@ -1397,7 +1397,7 @@ class ModelSet:
             line_center_conv = self.line_center.to(spec_unit).value
             vx = (spec - line_center_conv) / line_center_conv * c_km_s
 
-        cube_final = jnp.zeros((nspec, ny_sky_samp, nx_sky_samp))
+        cube_final = jnp.zeros((nspec, ny_sky_samp, nx_sky_samp), dtype=jnp.float64)
 
 
         # First construct the cube based on mass components
